requirements.txt
django-environ==0.9.0
Django==4.1.0
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
factory_boy==3.2.1
Faker==17.6.0
Markdown==3.3.3
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
requests==2.24.0
pytz==2023.3
